refactor(runtime): split types/binding.rs into cohesive submodules (§7-8)
Split the 1822-line runtime/types/binding.rs (over the 500-line limit) into
two cohesive sibling modules under src/runtime/types/. All methods remain in
impl Interpreter blocks; only four private helpers are bumped to pub(crate).

- binding_helpers.rs (206)   - named_scalar_container_share_eligible,
                               eval_param_default, implicit_method_named_slurpy,
                               check_shape_constraint
- binding_signature.rs (1619) - legacy_has_plain_positional_param (free fn),
                                parameter_binding_error,
                                normalize_coercion_binding_error,
                                bind_function_args_values (1568-line indivisible
                                giant; cannot be split at a function boundary)

Pure relocation. Four private methods bumped to pub(crate) to allow calls from
binding_signature.rs. The fn multiset is byte-identical. No behavior change.
